refactor: 콘솔 게임 실행 흐름 객체를 JanggiGameRunner로 변경
