+++ src/main/java/janggi/Application.java
@@ -1,12 +1,12 @@
 package janggi;
 
-import janggi.controller.JanggiController;
+import janggi.runner.JanggiGameRunner;
 import janggi.view.InputView;
 import janggi.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
-        JanggiController janggiController = new JanggiController(new InputView(), new OutputView());
-        janggiController.start();
+        JanggiGameRunner janggiGameRunner = new JanggiGameRunner(new InputView(), new OutputView());
+        janggiGameRunner.play();
     }
 }
